requirements.txt
streamlit
facenet-pytorch
pillow
numpy
python-dotenv
mysql-connector-python
cryptography
